Set environment variables in post-order phase.
This fixes https://github.com/reasonml/dependency-env/issues/8 but also makes more sense. Your dependencies can set variables, and then you can have the final say on what those are. It shouldn't be the other way around, with you setting environment variables, and your dependencies having the final say.

It still is not perfect, as there's no way for a dependency to say "I don't want anyone who depends on me to be able to reset this env var", but this is still an improvement.